feat: move packages from src to root of the project
