+++ miabis_model/collection_organization.py
@@ -6,12 +6,12 @@ from fhirclient.models.fhirreference import FHIRReference
 from fhirclient.models.meta import Meta
 from fhirclient.models.organization import Organization
 
-from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
-from src.MIABIS_on_FHIR.util.config import FHIRConfig
-from src.MIABIS_on_FHIR.util.constants import COLLECTION_DESIGN, COLLECTION_SAMPLE_COLLECTION_SETTING, \
+from miabis_model.incorrect_json_format import IncorrectJsonFormatException
+from miabis_model.util.config import FHIRConfig
+from miabis_model.util.constants import COLLECTION_DESIGN, COLLECTION_SAMPLE_COLLECTION_SETTING, \
     COLLECTION_SAMPLE_SOURCE, COLLECTION_DATASET_TYPE, COLLECTION_USE_AND_ACCESS_CONDITIONS
-from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value, parse_contact, parse_reference_id
-from src.MIABIS_on_FHIR.util.util import create_country_of_residence, create_contact, create_codeable_concept_extension, \
+from miabis_model.util.parsing_util import get_nested_value, parse_contact, parse_reference_id
+from miabis_model.util.util import create_country_of_residence, create_contact, create_codeable_concept_extension, \
     create_string_extension, create_fhir_identifier
