+++ miabis_model/network_organization.py
@@ -4,11 +4,11 @@ from fhirclient.models.fhirreference import FHIRReference
 from fhirclient.models.meta import Meta
 from fhirclient.models.organization import Organization
 
-from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
-from src.MIABIS_on_FHIR.util.config import FHIRConfig
-from src.MIABIS_on_FHIR.util.constants import NETWORK_COMMON_COLLAB_TOPICS
-from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value, parse_contact, parse_reference_id
-from src.MIABIS_on_FHIR.util.util import create_fhir_identifier, create_contact, create_country_of_residence, \
+from miabis_model.incorrect_json_format import IncorrectJsonFormatException
+from miabis_model.util.config import FHIRConfig
+from miabis_model.util.constants import NETWORK_COMMON_COLLAB_TOPICS
+from miabis_model.util.parsing_util import get_nested_value, parse_contact, parse_reference_id
+from miabis_model.util.util import create_fhir_identifier, create_contact, create_country_of_residence, \
     create_codeable_concept_extension, create_string_extension
