+++ miabis_model/observation.py
@@ -11,10 +11,10 @@ from fhirclient.models.fhirdatetime import FHIRDateTime
 from fhirclient.models.fhirreference import FHIRReference
 from fhirclient.models.meta import Meta
 
-from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
-from src.MIABIS_on_FHIR.util.config import FHIRConfig
-from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value, parse_reference_id
-from src.MIABIS_on_FHIR.util.util import create_fhir_identifier
+from miabis_model.incorrect_json_format import IncorrectJsonFormatException
+from miabis_model.util.config import FHIRConfig
+from miabis_model.util.parsing_util import get_nested_value, parse_reference_id
+from miabis_model.util.util import create_fhir_identifier
 
 
 class Observation:
