+++ pyproject.toml
@@ -4,8 +4,8 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "MIABIS_on_FHIR"
-version = "1.0.0"
-description = "Library containing classes for easier handling of data according to the MIABIS on FHIR profile"
+version = "1.1.0"
+description = "Library containing classes for easier handling of data according to the MIABIS on FHIR profile, as well as client for communication with sample blaze server"
 readme = "README.md"
 authors = [{"name" = "Simon Konar", "email" = "simon.konar@gmail.com"}]
 license = { file = "LICENSE" }
@@ -26,11 +26,8 @@ requires-python = ">=3.11"
 [project.optional-dependencies]
  test=["pytest >= 8.3.0"]
 
-[tool.setuptools]
-package-dir = { "" = "src" }
 
 
 [tool.setuptools.packages.find]
-where = ["src"]
-include = ["MIABIS_on_FHIR*", "blaze_client*"]
+include = ["miabis_model*", "blaze_client*"]
 exclude =["test*"]
