+++ test/unit/model/test_collection_org.py
@@ -1,6 +1,6 @@
 import unittest
 
-from src.MIABIS_on_FHIR.collection_organization import CollectionOrganization
+from miabis_model.collection_organization import CollectionOrganization
 
 
 class TestCollectionOrganization(unittest.TestCase):
